requirements.txt
python-dotenv==1.0.0
fastapi==0.111.1
fastapi-cli==0.0.4
uvicorn==0.30.3
supabase==2.4.0
pymongo==4.8.0